+++ CHANGELOG.md
@@ -1,4 +1,8 @@
+  - [AWS Route53 provider] Support customization of DNS record TTL through the use of annotation `external-dns.alpha.kubernetes.io/ttl` on services or ingresses (#320) @kevinjqiu
   - Added support for [DNSimple](https://dnsimple.com/) as DNS provider (#224) @jose5918
+
+## v0.4.5 - 2017-09-24
+
   - Add `--log-level` flag to control log verbosity and remove `--debug` flag in favour of `--log-level=debug` (#339) @ultimateboy
   - AWS: Allow filtering for private and public zones via `--aws-zone-type` flag (#329) @linki
   - CloudFlare: Add `--cloudflare-proxied` flag to toggle CloudFlare proxy feature (#340) @dunglas